fixes run.py and remove old bavaria scrapers
