+++ jedeschule/old/schule-gegen-rassisum.py
