+++ jedeschule/spiders/bayern.py
@@ -1,41 +1,43 @@
 # -*- coding: utf-8 -*-
 import scrapy
-from jedeschule.utils import cleanjoin
 from scrapy.shell import inspect_response
+
+from jedeschule.utils import get_first_or_none, cleanjoin
 
 
 class BayernSpider(scrapy.Spider):
     name = "bayern"
-    # allowed_domains = ["https://www.bllv.de/index.php?id=2707"]
-    start_urls = ['https://www.bllv.de/index.php?id=2707']
+    #allowed_domains = ["https://www.km.bayern.de/schueler/schulsuche.html"]
+    start_urls = ['https://www.km.bayern.de/schueler/schulsuche.html?s=&t=9999&r=9999&o=9999&u=0&m=3&seite=1']
 
     def parse(self, response):
-        school_types = response.css("#center_col > form > p:nth-child(2) > input ::attr(value)").extract()
-        print(school_types)
-        for school_type in school_types:
-            yield scrapy.FormRequest.from_response(
-                response,
-                callback=self.parse_list,
-                formdata={"varSchulart[]": school_type},
-            )
+        number_of_pages = response.css("div.schulsuche > div > p.Right a:last-child::text").extract_first()
+        # number_of_pages = 2
+        for i in range(1, int(number_of_pages) + 1):
+            url = "https://www.km.bayern.de/schueler/schulsuche.html?s=&t=9999&r=9999&o=9999&u=0&m=3&seite={page}"
+            yield scrapy.Request(url.format(page=i),
+                                 callback=self.parse_list)
 
     def parse_list(self, response):
-        school_numbers = response.css("table tr:nth-child(1) td:nth-child(2) ::text").extract()
-        for school_number in school_numbers:
-            yield scrapy.FormRequest.from_response(
-                response,
-                callback=self.parse_details,
-                formdata={"varSchulnummer": school_number},
-            )
+        links = response.css('.ListSchools a::attr(href)').extract()
+        for link in links:
+            yield scrapy.Request(response.urljoin(link), callback=self.parse_detail)
 
-    def parse_details(self, response):
+    def parse_detail(self, response):
+        # inspect_response(response, self)
+        text = response.css("article ::text")
         collection = {}
-        for tr in response.css("table tr"):
-            tds = tr.css("td ::text").extract()
-            # sometimes there is no value for the key
-            if len(tds) >= 2:
-                key = tds[0][:-1]
-                collection[key] = cleanjoin(tds[1:], "\n")
-                if key == "Name":
-                    collection['website'] = tr.css("td:nth-child(2) a::attr(href)").extract_first()
+        street, city = response.css("article > p")[0].css("::text").extract()
+        collection['street'] = street
+        collection['city'] = city
+        collection['name'] = cleanjoin(response.css('article h1::text').extract(), "")
+        collection['phone'] = get_first_or_none(text.re("Telefon: ([0-9 /]+)"))
+        collection['fax'] = get_first_or_none(text.re("Fax: ([0-9 /]+)"))
+        collection['web'] = response.css("article a::attr(href)").extract_first()
+        collection['number'] = get_first_or_none(text.re("Schulnummer: ([0-9]+)"))
+        collection['school_type'] = get_first_or_none(text.re("Schulart: (.+)"))
+        collection['type'] = get_first_or_none(text.re("Rechtlicher Status: (.+)"))
+        collection['teachers'] = get_first_or_none(text.re("Hauptamtliche Lehrkräfte: ([0-9]+)"))
+        collection['students'] = get_first_or_none(text.re("Schüler: ([0-9]+)"))
+        collection['url'] = response.url
         yield collection
